Move General helper to index
